+++ src/table_client/table_client.c
@@ -186,6 +186,10 @@ static ret_t table_client_on_paint_self(widget_t* widget, canvas_t* c) {
   return RET_OK;
 }
 
+static ret_t table_client_layout_children(widget_t* widget) {
+  return table_client_ensure_children(widget);
+}
+
 static ret_t table_client_prepare_data(widget_t* widget) {
   table_client_t* table_client = TABLE_CLIENT(widget);
 
@@ -636,6 +640,7 @@ TK_DECL_VTABLE(table_client) = {.size = sizeof(table_client_t),
                                 .parent = TK_PARENT_VTABLE(widget),
                                 .create = table_client_create,
                                 .on_paint_self = table_client_on_paint_self,
+                                .on_layout_children = table_client_layout_children,
                                 .set_prop = table_client_set_prop,
                                 .get_prop = table_client_get_prop,
                                 .on_event = table_client_on_event,