+++ docs/changes.md
@@ -1,5 +1,8 @@
 # 最新动态
 
+2020/12/16
+  * 修复awtk-widget-table-view中widget\_move\_resize函数无效导致行重叠的问题(感谢雨欣提供补丁)。
+
 2020/12/15
   * 修复awtk-widget-table-view在designer中默认xml有误导致行错位的问题(感谢雨欣提供补丁)。